sever/requirements.txt
ultralytics
opencv-python
Pillow
numpy
accelerate

torch
transformers
sentence-transformers
bitsandbytes

fastapi
uvicorn
python-multipart
deep-translator
